Make verifications safe to use on staging
=== WebhookDB Tables

Verification code required a legit webhookdb connection
and tables to load models; this wasn't easy to just
skip the code path using config, since it's in models.

So use a mock database if webhookdb models are not enabled.

=== Disable Front on verification list

If Front is not configured, disable the buttons that work with
conversations. These codepaths call Front,
so better not to call them.

=== Redundant events

Also fix the bug where taking an action on the verification list
would result in the same browser window seeing the update.

Instead, the event token acts like a session token,
and events taken by a given session are not reported
to the subscriber for that session.

Add an api helper to enforce the presence of the Suma-Events-Token
header when making unsafe (POST, etc) requests.

=== Fixes

Fix api.js so it accepts all Axios params,
not just the body/query.