Replace non-structured vendor registrations
Remove VendorAccount#registered_with_vendor column
with a new VendorAccountRegistration model.

A single vendor account can be registered with many vendor programs,
as is the case with Lyft Pass. We may as well standardize this,
since the 'one program' case is easy to work with while
support ing the 'many programs' case; while the inverse is not true,
we were having to store and parse JSON in a text field.