Member ordering (#832)
* Add ordering to associations

Let's try to avoid arbitrary ordering on most associations.
The performance cost is negligible for our use cases
(the set of rows we're sorting is very small due to FK indices).

* Add ordering to many-to-many associations

* All many_to_one associations must have FKs

Add a test, and migrate missing columns.
It is safer to assume we should always have these,
than try to remember to add them.
Associations can opt-out as needed.