Enrollment removal: Lyft handles deleted members
If a member is deleted, we error trying to revoke their lyft pass.
This fixes that bug by revoking the most recent previous phone
if the member is deleted.

Note that we do NOT process unenrollments when a member is
soft deleted. This was pretty confusing to set up,
and I'm not totally sure we want it yet,
so we will only still remove someone
when they're unenrolled
(we may want to remove all enrollments when someone is soft deleted).