Hybrid search conditional update needs to be in the DB
We cannot check clientside timestamps, since we could be
wiping out entire updates by other processes that ran within
our own timeframe (started and finished between when we fetch and
when we update).

If we do a conditional update using the DB (UPDATE .. WHERE ..),
we can only update things that are older and won't miss updates.