Fix race condition in hybrid search indexing (#856)
* Fix race condition in hybrid search indexing

It was possible that model reindexing could run out of order:
- Process A gets an update
- Process A starts indexing, calls the AI API
- Process B gets an update
- Process B starts indexing, calls the AI API
- Process B returns and updates embeddings with new data
- Process A returns and updates embeddings with old data

Instead, we now make sure that when Process A updates,
the hash column's internal timestamp if before the new timestamp.

This required adding a 'schema' to the hash column,
which is now 'v2'. We can use `<version>.<llm + text md5>.<timestamp>`
as the format.

* Hybrid search conditional update needs to be in the DB

We cannot check clientside timestamps, since we could be
wiping out entire updates by other processes that ran within
our own timeframe (started and finished between when we fetch and
when we update).

If we do a conditional update using the DB (UPDATE .. WHERE ..),
we can only update things that are older and won't miss updates.