Vendor service pricing is program-based
Previously, vendor services could point to a rate.
But because different members can have different rates,
the pricing needs to be done through the program.

This adds a new Program::Pricing model,
which allows a program to associate with a vendor service and rate.
Each program can only associate with a vendor service once,
so it can only have the same rate; the same service
can be on different programs though.

If a user has access to the same service through
multiple programs, it will use the rate with the lowest ordinal.

This modified some of the original model code
that predated even eligibility constraints,
so there was some remodeling required to make it work
across a bunch of files.

Thankfully most of the work is on the backend
(and new models in admin), only the mobility map on the web was
impacted, and even then it was able to be done without
changing the user-facing response shapes.