Overhaul how vendor services are priced (#864)
Vendor services and rates are an original model, that modeled vendor services as having many (really one) rate for a service. However, different members can have different prices for a service. This approach doesn't work anymore. Pricing needs to be done through programs (as access is done).

We added vendor service and rates to programs, but this only worked for lyft pass, where we had a program and could then get the service and rate from it. It doesn't work where we don't have an explicit program to work from, like when creating trips from Lime receipts (these come to an email, not a specific program). Instead, to get the lime service and rate, we need to find the service and rate available to the user when the trip was taken.

This adds a new 'program pricing' model that brings together programs (access), services, and rates. Each service can only be priced once in a program; a member could have access to a service through multiple programs, in which case we resolve the preferred rate by its ordinal.

This PR includes other supporting cleanup:

- Localization key for rates makes fewer assumptions- use any value within the 'strings' namespace.
- Upgrade tstzrange plugin to handle nil ranges
- ResourceDetail backTo supports string/func arg
- Support full modification of vendor services and rates in admin