Add Payment Instrument model
Payment instrument searching was a pain and is currently
almost entirely broken.

We now have a view-based `payment_instruments_view`,
which unions bank accounts and cards.
These tables are also now indexed using a hybrid search;
this search information is carried through to the view,
so we can do a normal hybrid search on the view.

This also reorganizes some instrument types:

- `Institution` moved from `Instrument` to `Payment`
- The module/mixin interface on `Instrument`
  moved to `Instrument::Interface`.