Further unification of instrument interface
We had some weird things, like brand repeating in labels.
Add more and better columns to the view so we can
fully represent cards and bank accounts.

This also eliminates 'now' from Ruby instrument work;
it was too error prone, and also made it impractical
to depend on database-derived data.