Hybrid search: Improve memory usage
Make vector columns load lazily instead of ahead of time,
causing less bloat in row size.
This has probably been the cause of memory bloat in prod.

This also exposed a bug with the translated text
`translation_join` helper, with how columns were selected
(well, it's more like a limitation with the column selection code,
since it doesn't handle join table aliases).
This fixes the code to use the table name as the join alias-
it could create new issues but hopefully not.