Add service/rate specific usage prohibition
In addition to global restrictions on service access
due to ledger balance,
we also want to prohibit using a service
if the rate is more than $0 and there is no default card.
This is because we know the user will be charged if
the service is used.

Show this better when selecting a vehicle on the map.