Resources with no programs are inaccessible
Previously, resources that could be linked to programs
(offerings, services) were available to everyone
if they were not in a program.
This was good for testing, but bad for production.

This changes the default so that:

- In testing, we preserve the previous behavior,
  since otherwise testing becomes a considerable pain
  (programs need to be created in dozens of tests,
  and it's easy to forget).
- Outside of testing, we do not include unlinked resources.

This should not introduce accidental issues
(or at least, no more than the current setup),
since this change does not end up provoking different
code paths in any caller; they just need to use `eligible_to`,
and everything will work right as for how the data
is set up.