Rewrite how mobility adapters work
Vendor services could point to a 'vendor adapter key',
like 'lime_maas' or 'lime_deeplink'.
These would effectively hard-code the vendors being used
to slugs configured for lime/lyft.

This makes things almost impossible to test effectively
since we cannot have lime/lyft outside of production.

Instead, Mobility::VendorAdapter is now a database object,
which is associated with a Vendor::Service (and thus a vendor),
and can specify if it uses deep linking or a mobility-as-a-service,
and whether suma or the vendor sends receipts.

Bootstrapped environments now do not need to specify lime or lyft;
instead we use analagous 'lemon' and 'rayse' vendors and services.