Fix 'Message us' lost account link in built app.
The `immediate` param didn't work right in the built app.

Using `navigate(/app/xyz)` was broken on the built app,
hosted at `/app`. We need to use `navigate(/xyz)` instead.

So use the `to` passed in (relative to `/app`, so `/xyz`),
rather than the anchor element href attr value
(which is resolved to `/app/xyz`).