AnonProxy: Handle users with disabled services
We do not want users to be able to use the API
to auth to a vendor if service use is prohibited.
This could happen for legit reasons, like the UI
allows the user to request a code, but it's now prohibited.

We also do not want to send requested access codes
onto the user if usage is prohibited.
This will generally only happen in malicious situations
(users knowing they are logged out, knowing their anonymous email,
and requesting the code through the Lime app).
So alert through Sentry for further manual action
(alerting support, banning, etc).