Rewrite trip syncs; abstract charging
This commit rewrites trip syncing to use the new trip design,
such that it works in a way much closer to commerce charging.

There is now a `Charge::Charger` abstract class
which can be overridden for things that need to charge,
like orders and trips.

This is now used during trip charging,
which removes the custom charging code,
and also adds proper subsidy matching.

Remove the lime email trip sync entirely for now.
We aren't using it so it isn't worth it to update,
and by the time we do use it,
it may be different.
I also have concerns about correctness,
especially with things like UI localization
(the receipts may not be in English).