Support first-class 'service revocation'
The enrollment remover had the ability to remove
a member from lime/lyft. But this is a more general capability
we need, like when a user loses access to services due to
nonpayment. Instead, create a new 'service revoker' concept.
This is called by enrollment remover
(which is good- it simplifies enrollment remover into just
dealing with enrollments, not the nuances of service removal).

It is also called:
- whenever a book transaction is created, in case the member
  is now unable to access services,
- every 30 minutes, with all potential candidates.

Service revocation can be in a 'dry run' mode,
which will allow us to run revocations
without potentially problematic side effects
if the selection criteria is bad.

This also creates a new view for ledger balances,
and a 'Ledger::Balance' model,
to make selecting potential service revocation candidates easier.