Messages ignore inaccessible vendor configuration
Do not handle messages we get where the user is not eligble.
Users can often request auth messages themselves, like a forgot password/magic link
using their anonymous contact info, in which case it'd come to us.
We want to ignore it; not send it onto the user.
NOTE: There is similar code in the lime handler;
maybe that should be generalized too.