Payment triggers can act as credits
Triggers have an `act_as_credit` field.
If true, they ignore the match multiplier behavior,
instead adding subsidy even for $0 in cash contributions.

This allows suma to cover the full cost of an item;
for example, to show Lyft users that their trips are not free,
but are fully subsidized by suma.