Admin: Close offering product
We have a way to change prices but not 'expire' an offering product.

This adds an endpoint and hooks up the 'delete' icon to close
the offering product.
Also displays closed_at in the UI.

Add an 'apiSoftDelete' to resource detail grid
so it will 'reload' the current grid rather than trying
to redirect to the page we just edited.
Add docs to ResourceDetail too.

Fixed wrong resource name in the anon member contact tables.